update GoodGhosting category for valora bakeoff (#128)
Hello,
We'd like to check if it's possible to update the category that GoodGhosting is listed in the Valora dApp for the bakeoff.
After checking the description of the category displayed on Valora app, seems like GoodGhosting is a better fit to the "earn" (lend) category, rather than "games". We'd appreciate if it's possible to get it updated. Thanks a lot.